Allow evaluation of external predictions
External predictions, if stored in PTB format, can now be
evaluated with the same code as the RNTN.

This commit refactors most of the original code into an
abstract class, AbstractEvaluate. Evaluate.java implements
the original functionality, so the change is fully compatible.
ExternalEvaluate.java implements the new functionality.

I release the changes proposed in this commit to the public domain
in accordance with the CONTRIBUTING.md document. --Michael Haas