NERCombinerAnnotator: fix IOOBE for empty text
For annotated empty text, the token list is empty. Therefore, assuming that there is at least one token in the list leads to exceptions. This assumption is made in the recently added transferNERAnnotationsToAnnotation method of NERCombinerAnnotator.

These changes check whether the token list is empty, and if it is then transferNERAnnotationsToAnnotation does nothing.