Remove an optimization in semgrex which may be causing crashes
The optimization removed - if it even does provide a noticeable time
benefit - can have repeated IdentityHashMap hashCodes for
SemanticGraphs over a long run, leading to crashes when a stale object
is retrieved for a new graph.

The directed graph now throws a specific exception type when a cyclic
graph occurs.  Catch exactly that exception in SemgrexMatcher
(previous versions caught Exception, which is horrifying)