Add an mwt misc field to the wire format, including the protobuf annotation serializer
Uses a single string instead of a map for mwtMisc.  This is not ideal,
because it would be nice to search through the individual pieces, but
different treebanks use different standards for how to separate the
pieces, so we just follow the same convention as for words for now and
keep the whole thing as one string.