Add 'uniq' as a keyword to SemgrexParser.jj Add a UniqPattern which removes duplicates based on the node names given (using the values of those nodes)
Add some unit testing of the uniq search parsing functionality -
SemgrexParser should fail if a requested node is not in the pattern.
Also, uniq should be usable as a node name

Test a couple varieties of this operation

To allow uniq for a ProcessSemgrexRequest, need to decode all sentences from the request first, then turn that into a response.
Flip the order of matching in ProcessSemgrexRequest so that for each pattern, it matches all of the sentences at once.  Allows for operations on the complete batch of matches, such as the new uniq operator

We also refactor the ProcessSemgrexRequest and make the CoreNLPServer use the refactored method as well

Add a test of uniq for ProcessSemgrexRequest as well (it should only produce one result now for two graphs, not two)